feat: Download datasets for training neural network
The MNIST dataset contains images of handwriting that will be used to train the neural network. Several lines have been added to the script to check if the dataset has been installed in a directory whose PATH is specified in a command line argument and to download the dataset if the dataset is missing. A command line argument parser has been added and the script must be called with a command line argument specifying the PATH or relative PATH to where the dataset is already downloaded or where the user wants to download the file. To run the main.py file in a development environment such as PyCharm, the configuration of the environment must be set to use the desired PATH as a command line argument.