Add functions for training the neural network
Using functions that utilize cross-entropy loss, the neural
network can now be trained by using the downloaded MNIST
dataset. The trained AI is then serialized and saved in a
.pt file. The script is able to run to completion, and
successfully saves the trained AI in the trained_AI.pt file,
but takes an incredibly long time to run. The next step is
to speed up the script with TorchPy's JIT