Reformat training duration log to use minutes and seconds
Rather than displaying the duration the training function
runs using the time in minutes up to some decimal point,
the reformatted log now displays the time in minutes and
seconds to make it easier for the user to grasp the time
interval. The docstring for nn_training was also extended.
It now clarifies why it is fine to simply call this function
without needing the function to return a value and
subsequently assign a variable to the return value in the
scope where nn_training is called.