Added CUDA functionality
To speed up operations that can be run in parallel, the main function will
now automatically switch to using the GPU if the computer running the
program has CUDA installed. This switch is automated and will not require
any command line input. Since the GPU has several hundred if not thousands
of cores compared to around a dozen with high-end Intel CPUs and several
dozen with high-end AMD CPUs (threadripper series). That, combined with the
fact that the OS can have well over a thousand threads being executed
on the CPU even when the computer is idling with no user applications
running, and switching to the GPU where possible using the CUDA
functionality should dramatically speed up the implementation on
machines with CUDA installed.